Some cleanup and norminette compliance
